+++ packages/common/src/models/Draw/DrawDataType.ts
@@ -1,4 +1,5 @@
-import { LotteryDrawData, RegularDrawData } from '../';
+import { LotteryDrawData } from '../Lottery';
+import { RegularDrawData } from '../Regular';
 
 /**
  * Type alias for types of draw data.