+++ packages/common/src/models/Game/index.ts
@@ -1,5 +1,4 @@
 export * from './Game';
-export * from './GameDrawDate';
 export * from './GameID';
 export * from './NumbersPool';
 export * from './Pool';