+++ packages/common/src/models/Match/MatchText.ts
@@ -2,5 +2,3 @@ export interface MatchText {
   long: string;
   short?: string;
 }
-
-export type MatchTextKeys = keyof MatchText;