+++ packages/common/src/models/Report/ReportLineType.ts
@@ -1,3 +1,4 @@
-import { LotteryReportLine, RegularReportLine } from './';
+import { RegularReportLine } from './RegularReportLine';
+import { LotteryReportLine } from './LotteryReportLine';
 
 export type ReportLineType = RegularReportLine | LotteryReportLine;