+++ packages/common/src/models/Ticket/Column.ts
@@ -1,6 +1,0 @@
-export interface Column {
-  main: number[];
-  plus?: number[];
-}
-
-export type ColumnKeys = keyof Column;