+++ packages/common/src/models/Ticket/Selection.ts
@@ -1,0 +1,7 @@
+/**
+ * Player's selection of numbers.
+ */
+export interface Selection {
+  main: number[];
+  plus?: number[];
+}