+++ packages/common/src/models/Ticket/index.ts
@@ -1,2 +1,2 @@
-export * from './Column';
+export * from './Selection';
 export * from './Ticket';